Further Consts progress
Added some more constants and the fields for the details window.
Window should be a JPanel as using JInternalFrame requires the whole main window to be changed into a JDesktop and using another JFrame is not optimal due to the lack of connectivity.
Using a custom JPanel with integrated functionality for minimizing, moving and scaling is the best approach.